refactor(people): one paging shape for the three lists, and a filter name spelled once
Fallout from the enumeration work, kept out of the commits that earned it.

The three list arms of SearchEntity differed only in the row type and where its
id and version sit, and adding a filter step to each pushed the function past
the complexity ceiling. The shared half is written once now: a per-type copy is
how one of them comes to page differently from its siblings without anyone
noticing.

A binding no longer repeats its own name. The set knows the key, so the FILTER
is named where the key is and the SHAPE where the parse is, which removes the
duplicate literals and leaves each name spelled once per module.
